feat: add OKLAB color space utilities and refactor vectorize to use them
